backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
autopep8==1.6.0
bcrypt==3.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.8
click==8.0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-contrib-postgres==0.0.1
django-cors-headers==3.10.1
django-environ==0.8.1
django-etc==1.3.1
django-hitcount==1.3.5
django-ipware==4.0.2
django-js-asset==2.0.0
django-oauth-toolkit==2.0.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-social-oauth2==1.2.1
ecdsa==0.17.0
factory-boy==3.2.1
Faker==13.3.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.3.1
MarkupSafe==2.0.1
msgpack==1.0.3
msgpack-python==0.5.6
mypy-extensions==0.4.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
pep8==1.7.1
Pillow==8.4.0
platformdirs==2.4.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
redis==2.10.6
requests==2.26.0
requests-oauthlib==1.3.1
rest-framework-generic-relations==2.1.0
rsa==4.8
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.14.1
zope.interface==5.4.0
